chore: refactor ClockView intro smaller files
