+++ meteor/client/ui/PieceIcons/PieceIcon.tsx
@@ -1,9 +1,7 @@
 import { withTracker } from '../../lib/ReactMeteorData/ReactMeteorData'
 import { MeteorReactComponent } from '../../lib/MeteorReactComponent'
 import * as React from 'react'
-import { Piece, Pieces } from '../../../lib/collections/Pieces'
 import { SourceLayerType, ISourceLayer, CameraContent, RemoteContent } from 'tv-automation-sofie-blueprints-integration'
-import { normalizeArray } from '../../../lib/lib'
 import * as _ from 'underscore'
 
 import CamInputIcon from './Renderers/CamInput'
@@ -12,93 +10,18 @@ import SplitInputIcon from './Renderers/SplitInput'
 import RemoteInputIcon from './Renderers/RemoteInput'
 import LiveSpeakInputIcon from './Renderers/LiveSpeakInput'
 import GraphicsInputIcon from './Renderers/GraphicsInput'
-import { Meteor } from 'meteor/meteor'
-import { ShowStyleBases, ShowStyleBaseId } from '../../../lib/collections/ShowStyleBases'
+import { ShowStyleBaseId } from '../../../lib/collections/ShowStyleBases'
 import { PubSub } from '../../../lib/api/pubsub'
-import { PieceInstances, PieceInstance } from '../../../lib/collections/PieceInstances'
+import { PieceInstance } from '../../../lib/collections/PieceInstances'
 import { PartInstanceId } from '../../../lib/collections/PartInstances'
 import { RundownId } from '../../../lib/collections/Rundowns'
+import { findPieceInstanceToShow } from './utils'
 
-interface IPropsHeader {
+export interface IPropsHeader {
 	partInstanceId: PartInstanceId
 	rundownIds: RundownId[]
 	showStyleBaseId: ShowStyleBaseId
 }
-
-interface INamePropsHeader extends IPropsHeader {
-	partName: string
-}
-
-function findPieceInstanceToShow(props: IPropsHeader, supportedLayers: Set<SourceLayerType>) {
-	let pieceInstances = PieceInstances.find({ partInstanceId: props.partInstanceId }).fetch()
-	let showStyleBase = ShowStyleBases.findOne(props.showStyleBaseId)
-
-	let sourceLayers = showStyleBase
-		? normalizeArray<ISourceLayer>(
-				showStyleBase.sourceLayers.map((layer) => ({ ...layer })),
-				'_id'
-		  )
-		: {}
-	let foundSourceLayer: ISourceLayer | undefined
-	let foundPiece: PieceInstance | undefined
-
-	for (const pieceInstance of pieceInstances) {
-		let layer = sourceLayers[pieceInstance.piece.sourceLayerId]
-		if (layer && layer.onPresenterScreen && supportedLayers.has(layer.type)) {
-			if (foundSourceLayer && foundPiece) {
-				if (
-					pieceInstance.piece.enable &&
-					foundPiece.piece.enable &&
-					((pieceInstance.piece.enable.start || 0) > (foundPiece.piece.enable.start || 0) ||
-						layer._rank >= foundSourceLayer._rank) // TODO: look into this, what should the do, really?
-				) {
-					console.log('Later: ', PieceIconContainer)
-					foundSourceLayer = layer
-					foundPiece = pieceInstance
-				}
-			} else {
-				console.log('First: ', pieceInstance)
-				foundSourceLayer = layer
-				foundPiece = pieceInstance
-			}
-		}
-	}
-
-	return {
-		sourceLayer: foundSourceLayer,
-		pieceInstance: foundPiece,
-	}
-}
-
-export const PieceNameContainer = withTracker((props: INamePropsHeader) => {
-	const supportedLayers = new Set([SourceLayerType.GRAPHICS, SourceLayerType.LIVE_SPEAK, SourceLayerType.VT])
-	return findPieceInstanceToShow(props, supportedLayers)
-})(
-	class PieceNameContainer extends MeteorReactComponent<
-		INamePropsHeader & { sourceLayer: ISourceLayer; pieceInstance: PieceInstance }
-	> {
-		componentDidMount() {
-			this.subscribe(PubSub.pieceInstancesSimple, {
-				rundownId: { $in: this.props.rundownIds },
-			})
-			this.subscribe(PubSub.showStyleBases, {
-				_id: this.props.showStyleBaseId,
-			})
-		}
-
-		render() {
-			if (this.props.sourceLayer) {
-				switch (this.props.sourceLayer.type) {
-					case SourceLayerType.GRAPHICS:
-					case SourceLayerType.LIVE_SPEAK:
-					case SourceLayerType.VT:
-						return this.props.pieceInstance.piece.name
-				}
-			}
-			return this.props.partName || ''
-		}
-	}
-)
 
 export const PieceIconContainer = withTracker((props: IPropsHeader) => {
 	const supportedLayers = new Set([
