+++ meteor/client/ui/StudioScreenSaver/StudioScreenSaver.tsx
@@ -150,6 +150,13 @@ export const StudioScreenSaver = translateWithTracker((props: IProps) => {
 			window.removeEventListener('resize', this.measureElement)
 		}
 
+		private static defaultPosition = (size: { width?: number; height?: number }) => {
+			return {
+				x: (window.innerWidth - (size.width || 0)) / 2,
+				y: (window.innerHeight - (size.height || 0)) / 2,
+			}
+		}
+
 		/**
 		 * When the component is mounted, the contents of the info box changes or when the window is resized,
 		 * measure it's size and calculate pixel value representations for all values that are specified in
@@ -164,10 +171,7 @@ export const StudioScreenSaver = translateWithTracker((props: IProps) => {
 
 				// this is a first measure, place the info-box in the center of the screen
 				if (this.state.infoElementSize.height === undefined || this.state.infoElementSize.width === undefined) {
-					this.position = {
-						x: (window.innerWidth - infoElementSize.width) / 2,
-						y: (window.innerHeight - infoElementSize.height) / 2,
-					}
+					this.position = StudioScreenSaver.defaultPosition(infoElementSize)
 
 					infoElement.style.transform = `translate3d(${this.position.x}px, ${this.position.y}px, 0.1px)`
 				}
@@ -201,20 +205,22 @@ export const StudioScreenSaver = translateWithTracker((props: IProps) => {
 			let { targetSpeedVector } = this.state
 			let { x, y } = this.position
 			let speedVector = this.speedVector
+			const windowWidth = window.innerWidth
+			const windowHeight = window.innerHeight
 			if (infoElement && infoElementSize.width && infoElementSize.height) {
 				// If the info element is within the frame margin, and the target vector is pointing outwards,
 				// calculate a new target vector and calculate the step that should happen per 1ms to transition
 				// the speedVector to the target in VECTOR_TRANSITION_DURATION time.
 				if (
 					(x < this.FRAME_PIXEL_MARGIN[3] && targetSpeedVector[0] < 0) ||
-					(x + infoElementSize.width > window.innerWidth - this.FRAME_PIXEL_MARGIN[1] && targetSpeedVector[0] > 0)
+					(x + infoElementSize.width > windowWidth - this.FRAME_PIXEL_MARGIN[1] && targetSpeedVector[0] > 0)
 				) {
 					targetSpeedVector = [targetSpeedVector[0] * -1, targetSpeedVector[1]]
 					this.vectorTransitionFrameStepX = (targetSpeedVector[0] - speedVector[0]) / this.VECTOR_TRANSITION_DURATION
 				}
 				if (
 					(y < this.FRAME_PIXEL_MARGIN[0] && targetSpeedVector[1] < 0) ||
-					(y + infoElementSize.height > window.innerHeight - this.FRAME_PIXEL_MARGIN[2] && targetSpeedVector[1] > 0)
+					(y + infoElementSize.height > windowHeight - this.FRAME_PIXEL_MARGIN[2] && targetSpeedVector[1] > 0)
 				) {
 					targetSpeedVector = [targetSpeedVector[0], targetSpeedVector[1] * -1]
 					this.vectorTransitionFrameStepY = (targetSpeedVector[1] - speedVector[1]) / this.VECTOR_TRANSITION_DURATION
@@ -238,9 +244,20 @@ export const StudioScreenSaver = translateWithTracker((props: IProps) => {
 					const normalizer = Math.max(Math.abs(speedVector[0]), Math.abs(speedVector[1]))
 					speedVector = [speedVector[0] / normalizer, speedVector[1] / normalizer]
 				}
+
 				this.position = {
 					x: x + speedVector[0] * this.PIXEL_SPEED * (frameTime / 17),
 					y: y + speedVector[1] * this.PIXEL_SPEED * (frameTime / 17),
+				}
+
+				// if by any chance the infoElement ends up beyond the screen area, reposition it to the center
+				if (
+					this.position.x < 0 ||
+					this.position.y < 0 ||
+					this.position.x + infoElementSize.width > windowWidth ||
+					this.position.y + infoElementSize.height > windowHeight
+				) {
+					this.position = StudioScreenSaver.defaultPosition(infoElementSize)
 				}
 
 				infoElement.style.transform = `translate3d(${this.position.x}px, ${this.position.y}px, 0.1px)`
